refactor: Move getMaticPOSClient to utils
